Create dataPull.py
Code to pull data from Illinois Sunshine. Gets the committee and its assorted data as well as the lat/lon, receipt volume, and receipt counter by year. This data is written to a csv file that can be then used in CartoDB to plot the points. Some changes might be needed in terms of storing the receipts by year.